Update message describing your changes

+++ database/migrations/2025_03_09_154252_remove_capitaine_from_users_table.php
@@ -14,8 +14,11 @@ class RemoveCapitaineFromUsersTable extends Migration
     public function up()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('capitaine');
+            if (Schema::hasColumn('users', 'capitaine')) {
+                $table->dropColumn('capitaine');
+            }
         });
+        
     }
 
     /**
